Refactor SourcePopover in own component
